+++ src/lib/compiler/precompile/determineUsedAssets.ts
@@ -4,26 +4,33 @@ import { MAX_NESTED_SCRIPT_DEPTH } from "consts";
 import { eventHasArg } from "lib/helpers/eventSystem";
 import type {
   CustomEvent,
-  Scene,
+  FontData,
+  // Scene,
   SoundData,
   Variable,
 } from "shared/lib/entities/entitiesTypes";
 import { EVENT_SOUND_PLAY_EFFECT } from "consts";
 import { walkScenesScripts } from "shared/lib/scripts/walk";
+import { ScriptEventHandlers } from "lib/project/loadScriptEventHandlers";
+import type { ProjectData } from "store/features/project/projectActions";
+import keyBy from "lodash/keyBy";
 
 export const determineUsedAssets = ({
-  scenes,
-  variablesLookup,
-  soundsLookup,
+  projectData,
   customEventsLookup,
+  scriptEventHandlers,
 }: {
-  scenes: Scene[];
-  variablesLookup: Dictionary<Variable>;
-  soundsLookup: Dictionary<SoundData>;
+  projectData: ProjectData;
   customEventsLookup: Dictionary<CustomEvent>;
+  scriptEventHandlers: ScriptEventHandlers;
 }) => {
+  const variablesLookup = keyBy(projectData.variables, "id");
+  const soundsLookup = keyBy(projectData.sounds, "id");
+  const fontsLookup = keyBy(projectData.fonts, "id");
+
   const usedVariablesLookup: Dictionary<Variable> = {};
   const usedSoundsLookup: Dictionary<SoundData> = {};
+  const usedFontsLookup: Dictionary<FontData> = {};
 
   const addAssetById =
     <T>(assetLookup: Dictionary<T>, usedLookup: Dictionary<T>) =>
@@ -32,6 +39,7 @@ export const determineUsedAssets = ({
       if (asset && !usedLookup[id]) {
         usedLookup[id] = asset;
       }
+      return !!asset;
     };
 
   const addVariableById = (id: string) => {
@@ -50,6 +58,14 @@ export const determineUsedAssets = ({
   };
   const addSoundById = addAssetById(soundsLookup, usedSoundsLookup);
 
+  const addFontById = addAssetById(fontsLookup, usedFontsLookup);
+
+  const addFontsFromString = (s: string) => {
+    (s.match(/(!F:[0-9a-f-]+!)/g) || [])
+      .map((id) => id.substring(3).replace(/!$/, ""))
+      .forEach(addFontById);
+  };
+
   const addReferences = (
     references: Reference[],
     filterType: string,
@@ -63,8 +79,15 @@ export const determineUsedAssets = ({
     }
   };
 
+  // Add default font
+  if (!addFontById(projectData.settings.defaultFontId)) {
+    if (projectData.fonts.length > 0) {
+      addFontById(projectData.fonts[0].id);
+    }
+  }
+
   walkScenesScripts(
-    scenes,
+    projectData.scenes,
     {
       customEvents: {
         lookup: customEventsLookup,
@@ -84,11 +107,32 @@ export const determineUsedAssets = ({
         const type = (cmd.args?.type as string) ?? "";
         addSoundById(type);
       }
+
+      if (cmd.args) {
+        for (const key in cmd.args) {
+          const value = cmd.args[key];
+          const fieldType =
+            scriptEventHandlers[cmd.command]?.fieldsLookup[key]?.type;
+
+          if (fieldType === "textarea" && typeof value === "string") {
+            // String text fields
+            addFontsFromString(value);
+          } else if (fieldType === "textarea" && Array.isArray(value)) {
+            // Multi value text fields
+            for (const row of value) {
+              if (typeof row === "string") {
+                addFontsFromString(row);
+              }
+            }
+          }
+        }
+      }
     }
   );
 
   return {
-    usedVariablesLookup,
-    usedSoundsLookup,
+    referencedVariables: Object.values(usedVariablesLookup) as Variable[],
+    referencedSounds: Object.values(usedSoundsLookup) as SoundData[],
+    referencedFonts: Object.values(usedFontsLookup) as FontData[],
   };
 };